requirements.txt
# ===================================
# CORE DEPENDENCIES
# ===================================
streamlit==1.31.1
pandas==2.2.0
python-dotenv==1.0.1

# ===================================
# DATABASE
# ===================================
oracledb==2.0.1
sqlalchemy==2.0.25

# ===================================
# HTTP REQUESTS & API
# ===================================
requests==2.31.0

# ===================================
# IMAGE PROCESSING
# ===================================
Pillow==10.2.0

# ===================================
# EXCEL EXPORT
# ===================================
openpyxl==3.1.2
XlsxWriter==3.1.9

# ===================================
# OPTIONAL (Para melhorias futuras)
# ===================================
# pytz==2024.1  # Timezone support
# python-magic==0.4.27  # Detecção de tipo de arquivo
# watchdog==4.0.0  # Monitor de arquivos